Add support for specifying release tag for okgo checks
Add a new "release-tag" configuration key that can be
used to specify the newest Go release tag that should
be used when checking Go files. Useful if the check
tool is used on a project that is being built with a
version of Go that is older than that used by the tool.

Fixes #72