Add support for generating multiple artifacts per dist
Makes it possible to support the notion of classifiers (a single
distribution with multiple outputs of the same type). Updates
Publish code so that POM for dist is published once per set of
artifacts for a particular distribution.

Also fixes bug in "artifacts" where, if there were multiple
Dist configurations of the same type for a single product, only
the outputs for one of them would be reported.

Fixes #98